fix: move profiling out of the cli
